Fix headless booting with wrong limit
Set the option default to 0 instead of 4.
Check if the limit is 0 instead of 4 when deciding to boot with default values.
Headless should no longer boot with default core count limit when booting with the override value 4.
Also fixed value not correctly being limited to 4+, which made it possible to run the emulator with less than 4 cores, which is not supported by a lot of games.